requirements.txt
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
mysql-connector-python==8.0.33
origamibot==2.3.6
protobuf==3.20.3
pycryptodomex==3.18.0
pyTelegramBotAPI==4.12.0
requests==2.31.0
shortuuid==1.0.11
soupsieve==2.4.1
telebot==0.0.5
typing_extensions==4.7.1
urllib3==2.0.3
websockets==11.0.3
yt-dlp==2023.6.22
